refactor: centralize sample data arrays
